Remove self.max_Q call from gradient-related methods
The algorithm for asynchronous deep RL uses the target network
parameters to compute the max_Q. This is unlike how it was coded earlier,
which used the same parameters. Thus, the call to self.max_Q is removed,
and the argument is made to hold the Q value instead of the next state.
To use it, one has to first run TargetNetwork.max_Q and then pass the
result to this method.

Other changes include:
- Rename Agent.py to FunctionApproximator.py, to allow extension to other
	approximation classes
- Edit gitignore to include TF checkpoints
- Remove test cases