Algo_Draft consists of  a rough structure behind the asynchronous q learning method
Here, just one thread is used. Actions are chosen according to epsilon-greedy. Training is done on the target network based on the new state.

Yet to be added :
train,restore methods for copying policy weights to target
To train over multiple threads

And yet to be optimized into arrays and cleaned